Crate atms

Ad-Hoc Threshold MultiSignatures (ATMS) implementation using Boldyreva multi signature scheme as described in Section 5.2 of the Proof-of-Stake Sidechains by Gazi, Kiayias and Zindros. Elliptic curve cryptography, and basic signature procedures are performed using the blst library by supranational which implements BLS signatures over curve BLS12-381.

Modules

aggregation

Aggregation module, which contains the mechanisms to compute and verify the aggregate signatures and keys of an ATMS signature protocol.

multi_sig

Multi signature module, which contains the wrappers around blst to build Boldryeva multi signature scheme.

Structs

MerkleTree

Tree of hashes, providing a commitment of data and its ordering.

MerkleTreeCommitment

Tree of hashes, providing a commitment of data and its ordering.

Path

Path of hashes from root to leaf in a Merkle Tree. Contains all hashes on the path, and the index of the leaf. Used to verify the credentials of users and signatures.

Enums

AtmsError

Errors associated with Atms