Module atms::aggregation

Aggregation module, which contains the mechanisms to compute and verify the aggregate signatures and keys of an ATMS signature protocol.

Notation

An ATMS signature protocol relies on an underlying multi-signature. For sake of simplicity we omit the Setup, KeyGen, Sign and Verify algorithms from the underlying signature, and refer to the documentation available in the multi signatures implementation. Given a set $S$, denote by $\langle S\rangle$ a Merkle-tree commitment to the set $S$ created in some fixed deterministic way. We divide the signers into two groups, eligible and participants. The eligible signers, $Es$ with $|Es| = n$, are all those who are allowed to participate in the signature process. We use all these keys to generate the master ATMS key. The participating signers, $Ps$ with $|Ps| = d$, are all those who participate in a particular signing procedure. We need $d$ to be larger than the threshold. Note that $Ps \subseteq Es$. Denote with $vk_i$ for $i \in Es$ the verification keys of all participants, and $\sigma_i$ for $i\in Ps$ the signatures produced by the participating users.

Structs

AggregateSig

An ATMS aggregate signature, contains the multi signature aggregation of all participating signers and the public key of all non-participating signers, together with a proof of membership to the vector commitment of all eligible signers.

Avk

An ATMS aggregate key, Avk, contains a vector commitment of all eligible signers, and the aggregated key. Any third party with access to the public keys from all eligible signers can generate an aggregate key.

Registration

An ATMS registration, which contains the aggregate key of all eligible signers, the Merkle tree containing all nodes in the tree (including the leaves), and a hash map, specifying the position of each key in the merkle tree commitment.