Module atms::merkle_tree

Merkle tree module, exposing several ops on Merkle Trees.

Structs

BatchPath

Path for a batch of indices. The size of a batched path, $s$, depends on how the nodes are distributed among the leaves. It has size $h − \log 2 k \leq s \leq k(h − \log 2 k)$, with $h$ the height of the tree and $k$ the size of the batch. This is considerably better than the trivial $k \cdot h$ solution of appending $k$ paths.

MerkleTree

Tree of hashes, providing a commitment of data and its ordering.

MerkleTreeCommitment

Tree of hashes, providing a commitment of data and its ordering.

Path

Path of hashes from root to leaf in a Merkle Tree. Contains all hashes on the path, and the index of the leaf. Used to verify the credentials of users and signatures.