Crate atms

Ad-Hoc Threshold MultiSignatures (ATMS) implementation using Boldyreva multi signature scheme as described in Section 5.2 of the Proof-of-Stake Sidechains by Gazi, Kiayias and Zindros. Elliptic curve cryptography, and basic signature procedures are performed using the blst library by supranational which implements BLS signatures over curve BLS12-381.

Modules

aggregation

Aggregation module, which contains the mechanisms to compute and verify the aggregate signatures and keys of an ATMS signature protocol.

merkle_tree

Merkle tree module, exposing several ops on Merkle Trees.

multi_sig

Multi signature module, which contains the wrappers around blst to build Boldryeva multi signature scheme.

Enums

AtmsError

Errors associated with Atms