Struct atms::aggregation::AggregateSig

pub struct AggregateSig<H> where
    H: Digest + FixedOutput,  {
    pub keys: Vec<PublicKey>,
    pub batch_proof: Option<BatchPath<H>>,
    // some fields omitted
}

An ATMS aggregate signature, contains the multi signature aggregation of all participating signers and the public key of all non-participating signers, together with a proof of membership to the vector commitment of all eligible signers.

Fields
keys: Vec<PublicKey>

Non-signing keys

batch_proof: Option<BatchPath<H>>

Batch proof of membership of non-signing keys

Implementations

impl<H> AggregateSig<H> where
H: Digest + FixedOutput,

pub fn new(
registration: &Registration<H>,
sigs: &[(usize, Signature)],
msg: &[u8]
) -> Result<Self, AtmsError>

Aggregate a list of signatures. The signature aggregation can again be performed by any third party. Given $d$ pairs of signatures, sigs, with their respective index representing the index in the merkle commitment, $\lbrace\sigma_i, id_i\rbrace_{i\in Ps}$, the aggregator produces the following aggregate signature. It begins by checking all signatures are valid, and which public keys are missing from the tuple of submitted signatures, $\widehat{vk}_i$, and computes their proof of set membership within the set of eligible signers, $\pi _{\widehat{vk_i}}$. Then it proceeds with the computation of the aggregate signature:

$$ \sigma = \left(\sigma_a = \sum_{i\in Ps}\sigma_i, \lbrace\widehat{vk}_i\rbrace _{i\in Es \setminus Ps }, \lbrace\pi _{\widehat{vk_i}}\rbrace _{i\in Es \setminus Ps}\right).$$

Error

Aggregation returns AtmsError::NonRegisteredParticipant if one of the submitted signatures comes from a non-registered participant, and AtmsError::InvalidSignature if one of the signatures is invalid.

pub fn verify(
&self,
msg: &[u8],
avk: &Avk<H>,
threshold: usize
) -> Result<(), AtmsError>

Verify that this aggregation, self, is a valid signature of msg with respect to the aggregate key avk with the given threshold. The verifier takes as input a message msg, an aggregate key avk, a signature $\sigma$ and a threshold $t$, and proceeds as follows:

Verify that all public keys are different, and that they belong to the commitment $\langle \mathcal{VK}\rangle$ in $avk$ using the proofs of membership.

Compute $avk’$ by dividing the aggregate key of non-signers, i.e. $$avk’ = avk - \sum_{i\in Es\setminus Ps}\widehat{vk_i}$$

Return valid if an only if $d\geq t$ and $\sigma$ validates with respect to $avk’$.

Error

Verification failds in the following cases:

AtmsError::FoundDuplicates if there are duplicates in the non-signers,

AtmsError::InvalidMerkleProof if the proof of membership is invalid, and

AtmsError::TooMuchOutstandingSigners if there are not enough signers.

Example

let n = 10; // number of parties
let subset_is = [1, 2, 3, 5, 6, 7, 9];
let threshold: usize = n - ((n - 1) / 3);
let msg = b"Did you know that Charles Babbage broke the Vigenere cipher?";
let mut rng = OsRng;

let mut sk_pks: Vec<(SigningKey, PublicKey)> = Vec::with_capacity(n);
let mut keyspop: Vec<PublicKeyPoP> = Vec::with_capacity(n);
let mut signatures: Vec<(usize, Signature)> = Vec::with_capacity(n);
for _ in 0..n {
    let sk = SigningKey::gen(&mut rng);
    let pk = PublicKey::from(&sk);
    let pkpop = PublicKeyPoP::from(&sk);
    keyspop.push(pkpop);
    sk_pks.push((sk, pk));
}

let atms_registration = Registration::<Blake2b>::new(&keyspop)?;

for i in 0..n {
    let (sk, pk) = &sk_pks[i];
    let sig = sk.sign(msg);
    assert!(sig.verify(pk, msg).is_ok());
    let indices = atms_registration.get_index(pk);
    for i in indices {
        signatures.push((i, sig));
    }
}
let avk = atms_registration.to_avk();
assert!(avk.check(&keyspop).is_ok());

let subset = subset_is
    .iter()
    .map(|i| {
        signatures[i % n]
    })
    .collect::<Vec<_>>();

let mut aggr_sig = AggregateSig::new(&atms_registration, &subset, msg).expect("Signatures should be valid");

aggr_sig.verify(msg, &avk, threshold).unwrap();

pub fn to_bytes(&self) -> Vec<u8>

Convert to a byte string of size $8 + b * 8 + 96 + t * (48 + P)$, where $b$ is boolean which is 1 if there are non-signers, $t$ is the number of non-signers, and $P$ is the size of the proof of membership.

Layout

The layout of an AggregateSignature is

Number of non-signers, $t$

Aggregate signature

$t$ public keys

Batch membership proof

pub fn from_bytes(bytes: &[u8]) -> Result<Self, AtmsError>

Deserialise a byte string to an AggregateSig.