Module atms::multi_sig

Multi signature module, which contains the wrappers around blst to build Boldryeva multi signature scheme.

Notation

We implement Boldryeva multi signatures over curve BLS12-381. Further reading can be found here and the references thereof cited. For level of detail required in this document, it is sufficient to understand that one can define a pairing over BLS12-381, where a pairing is a map: $e:\mathbb{G}_1 \times \mathbb{G}_2 \rightarrow \mathbb{G}_T$, which satisfies the following properties:

Bilinearity: $\forall a,b \in F^* _q$, $\forall P \in \mathbb{G}_1, Q \in \mathbb{G}_2: e(a* P,b* Q)=e(P,Q)^{ab}$

Non-degeneracy: $e \neq 1$

Computability: There exists an efficient algorithm to compute $e$

where $\mathbb{G}_1, \mathbb{G}_2$ and $\mathbb{G}_T$ are three distinct groups of order a prime $q$. We use additive notation for the operations over $\mathbb{G}_1, \mathbb{G}_2$ as the groups are defined in an elliptic curve, while we use multiplicative notation for $\mathbb{G}_T$ as the latter is defined in a multiplicative subgroup of an extension of $F_q$. We use $G_1, G_2$ to denote the generators of $\mathbb{G}_1$ and $\mathbb{G}_2$ respectively. Finally, we use a hash function $H_2: \lbrace 0,1\rbrace^* \rightarrow \mathbb{G}_2$ following the standardisation effort in the hashing to curves standard draft.

Structs

ProofOfPossession

Proof of possession, proving the correctness of a public key.

PublicKey

Public Key.

PublicKeyPoP

A public key with its proof of possession.

Signature

ATMS partial signature.

SigningKey

Signing key.